requirements.txt
#python=3.11.4
absl-py==1.4.0
accelerate==0.19.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.7.3
boto3==1.26.133
botocore==1.29.133
build==0.10.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
contourpy==1.0.7
crcmod==1.7
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.34
datasets==2.8.0
decorator==4.4.2
diffusers==0.16.1
dill==0.3.6
einops==0.6.1
executing==1.2.0
fairseq==0.12.2
filelock==3.12.0
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
ftfy==6.1.1
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
h5py==3.8.0
httplib2==0.22.0
huggingface-hub==0.14.1
hydra-core==1.0.7
idna==3.4
imageio==2.28.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
ipython==8.13.2
jax==0.4.10
jedi==0.18.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonplus==0.8.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
librosa==0.10.0.post2
lightning-utilities==0.8.0
llvmlite==0.40.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
modelscope==1.4.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
networkx==3.1
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.0.6
open-clip-torch==2.19.0
openai==0.27.6
opencv-python==4.7.0.72
opt-einsum==3.3.0
oss2==2.17.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pooch==1.6.0
portalocker==2.7.0
proglog==0.1.10
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.17
Pygments==2.15.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytorch-lightning==2.0.2
pytz==2023.3
pywin32==306
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
s3transfer==0.6.1
sacrebleu==2.3.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
simplejson==3.19.1
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
timm==0.9.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
xxhash==3.2.0
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
